+++ sdb_ai/sd64/gplbld/pcode_bld.py
@@ -114,7 +114,7 @@ def main():
     for src in pcode_fs:
         logger.info('**********************************************************************')
         result =   subprocess.run(
-            [sys.executable,bbcmp,SDSYS,'GPL.BP/'+src,'PCODE.OUT/'+src],
+            [sys.executable,bbcmp,SDSYS,'gpl.bp/'+src,'pcode.out/'+src],
              capture_output=True,
              text = True)
         logger.info(result.stdout)
@@ -130,11 +130,11 @@ def main():
             logger.info('**********************************************************************')
         #diff --suppress-common-lines <(xxd /usr/local/sdsys/GPL.BP.OUT/_HF) <(xxd /usr/local/sdsys/PCODE.OUT/_HF)    
             result =   subprocess.run(
-                ['xxd', SDSYS + '/GPL.BP.OUT/'+src, 'bsrc1'])
+                ['xxd', SDSYS + '/gpl.bp.out/'+src, 'bsrc1'])
     #             capture_output=True,
     #             text = True)
             result =   subprocess.run(
-                ['xxd', SDSYS + '/PCODE.OUT/'+src, 'bsrc2'])
+                ['xxd', SDSYS + '/pcode.out/'+src, 'bsrc2'])
     #             capture_output=True,
     #             text = True)
             
@@ -155,7 +155,7 @@ def main():
     p_idx = 0     # byte index for pcode_file
     for src in pcode_fs:
         f_byte_count = 0  # count of bytes in src file
-        fh = open(SDSYS + os.sep + 'PCODE.OUT' + os.sep + src, "rb")
+        fh = open(SDSYS + os.sep + 'pcode.out' + os.sep + src, "rb")
         filebytes = bytearray(fh.read())
         for fbyte in filebytes:
             pcode_file[p_idx] = fbyte